+++ Core/Gen3/EncounterArea3.hpp
@@ -28,7 +28,7 @@ class EncounterArea3 : public EncounterArea
 
 public:
     EncounterArea3() = default;
-    EncounterArea3(int location, u32 delay, Encounter type, const QVector<Slot> &pokemon);
+    EncounterArea3(u8 location, u32 delay, Encounter type, const QVector<Slot> &pokemon);
     u8 calcLevel(u8 index, u16 prng) const;
     u8 calcLevel(u8 index) const;
     u8 getEncounterRate() const;
